requirements.txt
numpy==1.24.4
opencv-python==4.5.2.52
Pillow
streamlit
torch==2.3.1
torchvision==0.18.1
scipy==1.10.1
